+++ springapp/src/main/java/com/example/springapp/repository/LoanRepository.java
@@ -5,7 +5,7 @@ import com.example.springapp.model.Loan;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoanRepository extends JpaRepository<Loan, Integer> {
+public interface LoanRepository extends JpaRepository<Loan, Long> {
     List<Loan> findByUserId(Long userId);
     List<Loan> findByApproverName(String approverName);
 }
