+++ springapp/src/main/java/com/example/springapp/service/LoanService.java
@@ -5,9 +5,9 @@ import java.util.List;
 import com.example.springapp.model.Loan;
 
 public interface LoanService {
-    Loan applyForLoan(String authHeader, int userId, Loan loan);
+    Loan applyForLoan(String authHeader, Long userId, Loan loan);
     List<Loan> getLoansByUserId(Long userId);
     List<Loan> getAllLoans();
-    Loan updateLoanStatus(int loanId, String status, String rejectReason);
+    Loan updateLoanStatus(Long loanId, String status, String rejectReason);
     List<Loan> getLoansByApproverName(String authHeader, String approverName);
 }
